newsboat: initial setup

+++ setup.sh
@@ -22,6 +22,7 @@ for name in \
     bat \
     emacs \
     git \
+    newsboat \
     nvim \
     ; \
 do
@@ -31,11 +32,13 @@ do
     fi
 done
 
-echo "symlink: ~/.config/alacritty/alacritty.local.yml"
-if [[ "$(uname)" == "Darwin" ]]; then
-    ln -s -f ~/.config/alacritty/alacritty.macos.yml ~/.config/alacritty/alacritty.local.yml
-else
-    ln -s -f ~/.config/alacritty/alacritty.linux.yml ~/.config/alacritty/alacritty.local.yml
+if [ ! -f ~/.config/alacritty/alacritty.local.yml ]; then
+    echo "symlink: ~/.config/alacritty/alacritty.local.yml"
+    if [[ "$(uname)" == "Darwin" ]]; then
+        ln -s -f ~/.config/alacritty/alacritty.macos.yml ~/.config/alacritty/alacritty.local.yml
+    else
+        ln -s -f ~/.config/alacritty/alacritty.linux.yml ~/.config/alacritty/alacritty.local.yml
+    fi
 fi
 
 paq_dir="${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/pack/paqs/start/paq-nvim
